third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGRectElement-svgdom-height-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' property of the SVGRectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rectElement.height.baseVal.value is 50
PASS rectElement.height.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
